chore: move AsciiWhitespaceSkippingReader to it's own module
