+++ src/main.rs
@@ -1,7 +1,7 @@
 use std::{
     collections::BTreeMap,
     env, fs,
-    io::{self, BufRead, BufReader, Read},
+    io::{self as std_io, BufReader, Read},
     str,
 };
 
@@ -13,9 +13,11 @@ use sha2::{Digest as _, Sha256, Sha512};
 use url::Url;
 
 use self::git::GitRepository;
+use self::io::AsciiWhitespaceSkippingReader;
 use self::registry::RegistryCrate;
 
 mod git;
+mod io;
 mod registry;
 mod rustup;
 
@@ -111,7 +113,7 @@ fn main() -> Result<()> {
         match package.checksum {
             Some(Checksum::Sha256(expected_sha256_hash)) => {
                 let mut sha256 = Sha256::new();
-                io::copy(&mut registry_crate.raw_crate_file()?, &mut sha256)?;
+                std_io::copy(&mut registry_crate.raw_crate_file()?, &mut sha256)?;
                 let sha256 = sha256.finalize();
 
                 ensure!(
@@ -283,10 +285,10 @@ fn main() -> Result<()> {
                 continue;
             }
 
-            let mut reader = AsciiWhitespaceSkippingReader(BufReader::new(file));
+            let mut reader = AsciiWhitespaceSkippingReader::new(BufReader::new(file));
 
             let mut sha512 = Sha512::new();
-            io::copy(&mut reader, &mut sha512)?;
+            std_io::copy(&mut reader, &mut sha512)?;
             crates_io_hashes.insert(path, sha512.finalize());
         }
 
@@ -302,10 +304,10 @@ fn main() -> Result<()> {
                 continue;
             }
 
-            let mut reader = AsciiWhitespaceSkippingReader(BufReader::new(file));
+            let mut reader = AsciiWhitespaceSkippingReader::new(BufReader::new(file));
 
             let mut sha512 = Sha512::new();
-            io::copy(&mut reader, &mut sha512)?;
+            std_io::copy(&mut reader, &mut sha512)?;
             our_hashes.insert(path, sha512.finalize());
         }
 
@@ -345,44 +347,4 @@ fn main() -> Result<()> {
     }
 
     Ok(())
-}
-
-struct AsciiWhitespaceSkippingReader<R>(R);
-
-impl<R> Read for AsciiWhitespaceSkippingReader<R>
-where
-    R: BufRead,
-{
-    fn read(&mut self, mut buf: &mut [u8]) -> io::Result<usize> {
-        let mut written = 0;
-
-        loop {
-            if buf.is_empty() {
-                break;
-            }
-
-            let mut read_buf = self.0.fill_buf()?;
-            if read_buf.is_empty() {
-                break;
-            }
-
-            let mut read = 0;
-            while !read_buf.is_empty() && !buf.is_empty() {
-                read += 1;
-                let b = read_buf[0];
-                read_buf = &read_buf[1..];
-                if b.is_ascii_whitespace() {
-                    continue;
-                }
-
-                buf[0] = b;
-                buf = &mut buf[1..];
-                written += 1;
-            }
-
-            self.0.consume(read);
-        }
-
-        Ok(written)
-    }
 }
